requirements.txt
Flask==1.1.2
pandas==1.2.1

requests==2.25.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2